+++ fasttree/HashTable.h
@@ -14,28 +14,28 @@ namespace fasttree {
         HashTable(const std::vector<std::string> &keys, bool checkUnique = false) {
             table.reserve(keys.size());
             if (checkUnique) {
-                for (size_t i = 0; i < keys.size(); i++) {
+                for (int64_t i = 0; i < (int64_t)keys.size(); i++) {
                     auto it = table.insert({&keys[i], i});
                     if (!it.second) {
                         throw std::invalid_argument("Non-unique name '" + keys[i] + "' in the alignment");
                     }
                 }
             } else {
-                for (size_t i = keys.size(); i > 0; i--) {
+                for (int64_t i = keys.size(); i > 0; i--) {
                     table.insert_or_assign(&keys[i - 1], i - 1);
                 }
             }
         }
 
-        inline size_t size() {
+        inline int64_t size() {
             return table.size();
         }
 
-        inline size_t operator[](const std::string &key) const {
+        inline int64_t operator[](const std::string &key) const {
             return table.find(&key)->second;
         }
 
-        inline const size_t *find(const std::string &key) const {
+        inline const int64_t *find(const std::string &key) const {
             auto entry = table.find(&key);
             if (entry != table.end()) {
                 return &(entry->second);
@@ -46,14 +46,14 @@ namespace fasttree {
     private:
 
         struct Hash {
-            size_t operator()(const std::string *p) const noexcept { return xxh64::hash(p->data(), p->size(), 0); }
+            int64_t operator()(const std::string *p) const noexcept { return xxh64::hash(p->data(), p->size(), 0); }
         };
 
         struct Equal {
             bool operator()(const std::string *x, const std::string *y) const { return *x == *y; }
         };
 
-        tsl::robin_map<const std::string *, size_t, Hash, Equal> table;
+        tsl::robin_map<const std::string *, int64_t, Hash, Equal> table;
     };
 }
