+++ fasttree/operations/SSEOperations.tcc
@@ -9,9 +9,9 @@ template<typename Precision> \
 __VA_ARGS__ fasttree::SSEOperations<Precision>
 
 template<>
-void fasttree::SSEOperations<float>::vector_add_mult(float fTot[], float fAdd[], float weight, size_t n) {
+void fasttree::SSEOperations<float>::vector_add_mult(float fTot[], float fAdd[], float weight, int64_t n) {
     __m128 w = _mm_set1_ps(weight);
-    for (size_t i = 0; i < n; i += 4) {
+    for (int64_t i = 0; i < n; i += 4) {
         __m128 tot, add;
         tot = _mm_load_ps(fTot + i);
         add = _mm_load_ps(fAdd + i);
@@ -20,9 +20,9 @@ void fasttree::SSEOperations<float>::vector_add_mult(float fTot[], float fAdd[],
 }
 
 template<>
-void fasttree::SSEOperations<double>::vector_add_mult(double fTot[], double fAdd[], double weight, size_t n) {
+void fasttree::SSEOperations<double>::vector_add_mult(double fTot[], double fAdd[], double weight, int64_t n) {
     __m128d w = _mm_set1_pd(weight);
-    for (size_t i = 0; i < n; i += 2) {
+    for (int64_t i = 0; i < n; i += 2) {
         __m128d tot, add;
         tot = _mm_load_pd(fTot + i);
         add = _mm_load_pd(fAdd + i);
